rag_retriever_service/requirements.txt
python-dotenv==1.0.1
llama-index>=0.9.31
chromadb==0.5.23
llama-index-vector-stores-chroma==0.4.1
sentence-transformers==3.3.1
llama-index-embeddings-huggingface==0.4.0
transformers==4.46.3
torch==2.2.2
numpy<2.0,==1.26.4
llama-index-llms-huggingface==0.4.1
jinja2==3.1.4
Hypercorn==0.17.3
pyjwt==2.10.1